requirements.txt
fastapi==0.88.0
gunicorn==20.1.0
pyjwt[crypto]==2.6.0
uvicorn[standard]==0.20.0
xlwings-0.0.0+2f7a8e5.tar.gz
